+++ src/lt_random.h
@@ -10,6 +10,8 @@
  * @license For the license see file LICENSE.txt file in the root directory of this source tree.
  */
 
+#include <stdlib.h>
+
 #include "libtropic_common.h"
 
 /**
@@ -19,6 +21,6 @@
  * @param count       Number of random bytes
  * @return lt_ret_t
  */
-lt_ret_t lt_random_bytes(void *buff, uint16_t count);
+lt_ret_t lt_random_bytes(void *buff, size_t count);
 
 #endif
